rf: extract validate_interval_format function
