+++ deps/wazuh_testing/wazuh_testing/tools/utils.py
@@ -4,6 +4,7 @@
 import logging
 import re
 import string
+import numbers
 from functools import wraps
 from random import randint, SystemRandom
 from time import sleep
@@ -153,3 +154,13 @@ def get_host_name():
         str: The host name.
     """
     return socket.gethostname()
+
+
+def validate_interval_format(interval):
+    """Validate that the interval passed has the format in which the last digit is a letter from those passed and
+       the other characters are between 0-9"""
+    if interval=='':
+        return False
+    if interval[-1] not in ['s','m', 'h','d','w','y'] or not isinstance(int(interval[0:-1]), numbers.Number):
+        return False
+    return True
